python/requirements.txt
easydict==1.6
PyYAML==3.11
wheel==0.24.0
